Use Catalogue of Life to get downstream taxa to a given taxonomic level.

Use Catalogue of Life to get downstream taxa to a given taxonomic level.

col_downstream(name = NULL, id = NULL, downto, format = NULL,
  start = NULL, checklist = NULL, verbose = TRUE, intermediate = FALSE,
  ...)

Arguments

name

The string to search for. Only exact matches found the name given will be returned, unless one or wildcards are included in the search string. An * (asterisk) character denotes a wildcard; a character may also be used. The name must be at least 3 characters long, not counting wildcard characters.

id

The record ID of the specific record to return (only for scientific names of species or infraspecific taxa)

downto

The taxonomic level you want to go down to. See examples below. The taxonomic level IS case sensitive, and you do have to spell it correctly. See data(rank_ref) for spelling.

format

The returned format (default = NULL). If NULL xml is used. Currently only xml is supported.

start

The first record to return (default = NULL). If NULL, the results are returned from the first record (start=0). This is useful if the total number of results is larger than the maximum number of results returned by a single Web service query (currently the maximum number of results returned by a single query is 500 for terse queries and 50 for full queries).

checklist

The year of the checklist to query, if you want a specific year's checklist instead of the lastest as default (numeric).

verbose

Print or suppress messages.

intermediate

(logical) If TRUE, return a list of length two with target taxon rank names, with additional list of data.frame's of intermediate taxonomic groups. Default: FALSE

...

Curl options passed on to GET

Value

A list of data.frame's.

Details

Provide only names instead of id's

Examples

## Not run: ------------------------------------
# # Some basic examples
# col_downstream(name="Apis", downto="species")
# col_downstream(name="Bryophyta", downto="family")
# 
# # get classes down from the kingdom Animalia
# col_downstream(name="Animalia", downto="class")
# col_downstream(name="Animalia", downto="class", intermediate=TRUE)
# 
# # An example that takes a bit longer
# col_downstream(name=c("Plantae", "Animalia"), downto="class")
# 
# # Using a checklist from a specific year
# col_downstream(name="Bryophyta", downto="family", checklist=2009)
# 
# # By id
# col_downstream(id='576d098d770a39d09e2bcfa1c0896b26', downto="species",
#   checklist=2012)
## ---------------------------------------------