EUBON hierarchy

EUBON hierarchy

eubon_hierarchy(id, providers = "pesi", timeout = 0, ...)

Arguments

id

(character) identifier for the taxon. (LSID, DOI, URI, or any other identifier used by the checklist provider)

providers

(character) A list of provider id strings concatenated by comma characters. The default : "pesi,bgbm-cdm-server[col]" will be used if this parameter is not set. A list of all available provider ids can be obtained from the '/capabilities' service end point. Providers can be nested, that is a parent provider can have sub providers. If the id of the parent provider is supplied all subproviders will be queried. The query can also be restriced to one or more subproviders by using the following syntax: parent-id[sub-id-1,sub-id2,...]

timeout

(numeric) The maximum of milliseconds to wait for responses from any of the providers. If the timeout is exceeded the service will just return the resonses that have been received so far. The default timeout is 0 ms (wait for ever)

...

Curl options passed on to GET

References

http://cybertaxonomy.eu/eu-bon/utis/1.2/doc.html

See also

Other eubon-methods: eubon_capabilities, eubon_children, eubon

Examples

## Not run: ------------------------------------
# eubon_hierarchy(id = "urn:lsid:marinespecies.org:taxname:126141", 'worms')
# eubon_hierarchy(id = "urn:lsid:marinespecies.org:taxname:274350", 'worms')
## ---------------------------------------------