Get a IUCN Redlist taxon

Get a IUCN Redlist taxon

get_iucn(sci, messages = TRUE, key = NULL, x = NULL, ...)

as.iucn(x, check = TRUE, key = NULL)

# S3 method for iucn
as.iucn(x, check = TRUE, key = NULL)

# S3 method for character
as.iucn(x, check = TRUE, key = NULL)

# S3 method for list
as.iucn(x, check = TRUE, key = NULL)

# S3 method for numeric
as.iucn(x, check = TRUE, key = NULL)

# S3 method for data.frame
as.iucn(x, check = TRUE, key = NULL)

# S3 method for iucn
as.data.frame(x, ...)

Arguments

sci

(character) A vector of scientific names. Or, a taxon_state object (see taxon-state)

messages

logical; should progress be printed?

key

(character) required. you IUCN Redlist API key. See rredlist::rredlist-package for help on authenticating with IUCN Redlist

x

For get_iucn(): Deprecated, see sci. For as.iucn(), various, see examples

...

Ignored

check

(logical) Check if ID matches any existing on the DB, only used in as.iucn()

Value

A vector of taxonomic identifiers as an S3 class.

Comes with the following attributes:

match (character) - the reason for NA, either 'not found', 'found' or if ask = FALSE then 'NA due to ask=FALSE')

name (character) - the taxonomic name, which is needed in synonyms() and sci2comm() methods since they internally use rredlist functions which require the taxonomic name, and not the taxonomic identifier

ri (character) - The URI where more information can be read on the taxon - includes the taxonomic identifier in the URL somewhere

multiple_matches and pattern_match do not apply here as in other get_*

methods since there is no IUCN Redlist search, so you either get a match or you do not get a match.

Details

There is no underscore method, because there's no real search for IUCN, that is, where you search for a string, and get back a bunch of results due to fuzzy matching. If that exists in the future we'll add an underscore method here.

IUCN ids only work with synonyms() and sci2comm() methods.

See also

Examples

if (FALSE) {
get_iucn("Branta canadensis")
get_iucn("Branta bernicla")
get_iucn("Panthera uncia")

# as coercion
as.iucn(22732)
as.iucn("22732")
(res <- as.iucn(c(22679946, 22732, 22679935)))
data.frame(res)
as.iucn(data.frame(res))
}