Get the uBio id for a search term

THIS FUNCTION IS DEFUNCT.

get_ubioid(searchterm, searchtype = "scientific", ask = TRUE,
  verbose = TRUE, rows = NA, family = NULL, rank = NULL, ...)

as.ubioid(x, check = TRUE)

# S3 method for ubioid
as.ubioid(x, check = TRUE)

# S3 method for character
as.ubioid(x, check = TRUE)

# S3 method for list
as.ubioid(x, check = TRUE)

# S3 method for numeric
as.ubioid(x, check = TRUE)

# S3 method for data.frame
as.ubioid(x, check = TRUE)

# S3 method for ubioid
as.data.frame(x, ...)

get_ubioid_(searchterm, verbose = TRUE, searchtype = "scientific",
  rows = NA)

Arguments

searchterm

character; A vector of common or scientific names.

searchtype

character; One of 'scientific' or 'common', or any unique abbreviation

ask

logical; should get_tsn be run in interactive mode? If TRUE and more than one TSN is found for teh species, the user is asked for input. If FALSE NA is returned for multiple matches.

verbose

logical; should progress be printed?

rows

numeric; Any number from 1 to infinity. If the default NA, all rows are considered. Note that this function still only gives back a ubioid class object with one to many identifiers. See get_ubioid_ to get back all, or a subset, of the raw data that you are presented during the ask process.

family

(character) A family name. Optional. See Filtering below.

rank

(character) A taxonomic rank name. See rank_ref for possible options. Though note that some data sources use atypical ranks, so inspect the data itself for options. Optional. See Filtering below.

...

Ignored

x

Input to as.ubioid

check

logical; Check if ID matches any existing on the DB, only used in as.ubioid

Value

A vector of uBio ids. If a taxon is not found NA is given. If more than one uBio id is found the function asks for user input (if ask = TRUE), otherwise returns NA. Comes with an attribute match to investigate the reason for NA (either 'not found', 'found' or if ask = FALSE 'NA due to ask=FALSE')

Filtering

The parameters family and rank are not used in the search to the data provider, but are used in filtering the data down to a subset that is closer to the target you want. For all these parameters, you can use regex strings since we use grep internally to match. Filtering narrows down to the set that matches your query, and removes the rest.

See also

get_uid, ubio_search