Search for taxonomic names using the Global Names Index.

Uses the Global Names Index, see http://gni.globalnames.org/.

gni_search(search_term = NULL, per_page = NULL, page = NULL,
  justtotal = FALSE, parse_names = FALSE, ...)

Arguments

search_term

Name pattern you want to search for. WARNING: Does not work for vernacular/common names. Search term may include following options (Note: can, uni, gen, sp, ssp, au, yr work only for parsed names):

* wild card - Search by part of a word (E.g.: planta*)

exact exact match - Search for exact match of a literal string (E.g.: exact:Parus major)

ns name string- Search for literal string from its beginning (other modifiers will be ignored) (E.g.: ns:parus maj*)

can canonical form- Search name without authors (other modifiers will be ignored) (E.g.: can:parus major)

uni uninomial- Search for higher taxa (E.g.: uni:parus)

gen genus - Search by genus epithet of species name (E.g.: gen:parus)

sp species - Search by species epithet (E.g.: sp:major)

ssp subspecies - Search by infraspecies epithet (E.g.: ssp:major)

au author - Search by author word (E.g.: au:Shipunov)

yr year - Search by year (E.g.: yr:2005)

per_page

Number of items per one page (numbers larger than 1000 will be decreased to 1000) (default is 30).

page

Page number you want to see (default is 1).

justtotal

Return only the total results found.

parse_names

If TRUE, use gni_parse to parse names. Default: FALSE

...

Curl options passed on to GET

Value

data.frame of results.

Details

Note that you can use fuzzy searching, e.g., by attaching an asterisk to the end of a search term. See the first two examples below.

References

http://gni.globalnames.org/, https://github.com/dimus/gni/wiki/api

See also

gnr_datasources, gni_search.

Examples

## Not run: ------------------------------------
# gni_search(search_term = "ani*")
# gni_search(search_term = "ama*", per_page = 3, page = 21)
# gni_search(search_term = "animalia", per_page = 8, page = 1)
# gni_search(search_term = "animalia", per_page = 8, page = 1, justtotal=TRUE)
# 
# gni_search(search_term = "Cyanistes caeruleus", parse_names=TRUE)
# 
# # pass on curl options to httr
# library("httr")
# gni_search(search_term = "ani*", config = verbose())
## ---------------------------------------------