Ping an API used in taxize to see if it's working.

Ping an API used in taxize to see if it's working.

col_ping(what = "status", ...)

eol_ping(what = "status", ...)

itis_ping(what = "status", ...)

ncbi_ping(what = "status", ...)

tropicos_ping(what = "status", ...)

nbn_ping(what = "status", ...)

gbif_ping(what = "status", ...)

bold_ping(what = "status", ...)

ipni_ping(what = "status", ...)

vascan_ping(what = "status", ...)

fg_ping(what = "status", ...)

Arguments

what

(character) One of status (default), content, or an HTTP status code. If status, we just check that the HTTP status code is 200, or similar signifying the service is up. If content, we do a simple, quick check to determine if returned content matches what's expected. If an HTTP status code, it must match an appropriate code. See status_codes.

...

Curl options passed on to GET

Value

A logical, TRUE or FALSE

Details

For ITIS, see description, which provides number of scientific and common names in a character string.

Examples

## Not run: ------------------------------------
# col_ping()
# col_ping("content")
# col_ping(200)
# col_ping("200")
# col_ping(204)
# 
# itis_ping()
# eol_ping()
# ncbi_ping()
# tropicos_ping()
# nbn_ping()
# 
# gbif_ping()
# gbif_ping(200)
# 
# bold_ping()
# bold_ping(200)
# bold_ping("content")
# 
# ipni_ping()
# ipni_ping(200)
# ipni_ping("content")
# 
# vascan_ping()
# vascan_ping(200)
# vascan_ping("content")
# 
# # curl options
# library("httr")
# vascan_ping(config=verbose())
# eol_ping(500, config=verbose())
## ---------------------------------------------