Species names from Species Plantarum

These names have been compiled from Species Plantarum by Carl Linnaeus originally published in 1753. It is the first work to consistently apply binomial names and was the starting point for the naming of plants. The book lists every species of plant known at the time, classified into genera. The dataset provides a useful reference point to see how taxonomic names have changed since their inception. The names were transcribed by Robert W. Kiger.

Format

A data frame with 5940 rows and 3 variables:

genus First part of the binomial species name for each species within the genus

epithet specific epithet or second part of the binomial species name for each species

page_number The following abbreviations sometimes are used in the page_number field.

"add." refers to addenda that appear on the unnumbered last page of the index in volume two.

"err." refers to the unnumbered page of errata that appears following the index in volume two.

"canc." following a page number indicates that the binomial appeared on the cancelled version of that page and does not appear on its replacement (as in the 1957-1959 facsimile edition.

Source

Hunt Institute for Botanical Documentation

References

Linnaeus, C. 1753. Species Plantarum. 2 vols. Salvius, Stockholm. [Facsimile edition, 1957-1959, Ray Society, London.]

Author

Carl Linnaeus