Get taxonomic names for a given rank

Get taxonomic names for a given rank

tax_name(query, get, db = "itis", pref = "ncbi", verbose = TRUE, ...)

Arguments

query

(character) Vector of taxonomic names to query. required.

get

(character) The ranks of the taxonomic name to get, see rank_ref. required.

db

(character) The database to search from: 'itis', 'ncbi' or 'both'. If 'both' both NCBI and ITIS will be queried. Result will be the union of both.

pref

(character) If db = 'both', sets the preference for the union. Either 'ncbi' (default) or 'itis'. Currently not implemented.

verbose

(logical) If TRUE the actual taxon queried is printed on the console.

...

Other arguments passed to get_tsn or get_uid.

Value

A data.frame with one column for every queried rank, in addition to a column for db and queried term.

Note

While tax_rank returns the actual rank of a taxon, tax_name searches and returns any specified rank higher in taxonmy.

See also

classification

Examples

## Not run: ------------------------------------
# # A case where itis and ncbi use the same names
# tax_name(query = "Helianthus annuus", get = "family", db = "itis")
# tax_name(query = "Helianthus annuus", get = "family", db = "ncbi")
# tax_name(query = "Helianthus annuus", get = c("genus","family","order"), db = "ncbi")
# 
# # Case where itis and ncbi use different names
# tax_name(query = "Helianthus annuus", get = "kingdom", db = "itis")
# tax_name(query = "Helianthus annuus", get = "kingdom", db = "ncbi")
# 
# # multiple rank arguments
# tax_name(query = c("Helianthus annuus","Baetis rhodani"), get = c("genus",
# "kingdom"), db = "ncbi")
# tax_name(query = c("Helianthus annuus","Baetis rhodani"), get = c("genus",
# "kingdom"), db = "itis")
# 
# # query both sources
# tax_name(query=c("Helianthus annuus", 'Baetis rhodani'), get=c("genus",
# "kingdom"), db="both")
## ---------------------------------------------