Retrieve immediate children taxa for a given taxon name or ID.

This function is different from downstream in that it only collects immediate taxonomic children, while downstream collects taxonomic names down to a specified taxonomic rank, e.g., getting all species in a family.

children(...)

# S3 method for default
children(x, db = NULL, rows = NaN, ...)

# S3 method for tsn
children(x, db = NULL, ...)

# S3 method for colid
children(x, db = NULL, ...)

# S3 method for wormsid
children(x, db = NULL, ...)

# S3 method for ids
children(x, db = NULL, ...)

# S3 method for uid
children(x, db = NULL, ...)

Arguments

...

Further args passed on to col_children, hierarchy_down, ncbi_children, or wm_children See those functions for what parameters can be passed on.

x

Vector of taxa names (character) or IDs (character or numeric) to query.

db

character; database to query. One or more of itis, col, ncbi, or worms. Note that each taxonomic data source has their own identifiers, so that if you provide the wrong db value for the identifier you could get a result, but it will likely be wrong (not what you were expecting).

rows

(numeric) Any number from 1 to infinity. If the default NA, all rows are considered. Note that this parameter is ignored if you pass in a taxonomic id of any of the acceptable classes: tsn, colid. NCBI has a method for this function but rows doesn't work.

Value

A named list of data.frames with the children names of every supplied taxa. You get an NA if there was no match in the database.

Examples

## Not run: ------------------------------------
# # Plug in taxonomic IDs
# children(161994, db = "itis")
# children(8028, db = "ncbi")
# children("578cbfd2674a9b589f19af71a33b89b6", db = "col")
# ## works with numeric if as character as well
# children("161994", db = "itis")
# 
# # Plug in taxon names
# children("Salmo", db = 'col')
# children("Salmo", db = 'itis')
# children("Salmo", db = 'ncbi')
# children("Salmo", db = 'worms')
# 
# # Plug in IDs
# (id <- get_colid("Apis"))
# children(id)
# 
# (id <- get_wormsid("Platanista"))
# children(id)
# 
# ## Equivalently, plug in the call to get the id via e.g., get_colid
# ## into children
# (id <- get_colid("Apis"))
# children(id)
# children(get_colid("Apis"))
# 
# # Many taxa
# sp <- c("Tragia", "Schistocarpha", "Encalypta")
# children(sp, db = 'col')
# children(sp, db = 'itis')
# 
# # Two data sources
# (ids <- get_ids("Apis", db = c('col','itis')))
# children(ids)
# ## same result
# children(get_ids("Apis", db = c('col','itis')))
# 
# # Use the rows parameter
# children("Poa", db = 'col')
# children("Poa", db = 'col', rows=1)
# 
# # use curl options
# library("httr")
# res <- children("Poa", db = 'col', rows=1, config=verbose())
# res <- children("Salmo", db = 'itis', config=verbose())
# res <- children("Salmo", db = 'ncbi', config=verbose())
## ---------------------------------------------