Search for pages in EOL database using a taxonconceptID.

Search for pages in EOL database using a taxonconceptID.

eol_pages(taxonconceptID, iucn = FALSE, images = 0, videos = 0,
  sounds = 0, maps = 0, text = 0, subjects = "overview",
  licenses = "all", details = FALSE, common_names = FALSE,
  synonyms = FALSE, references = FALSE, taxonomy = TRUE, vetted = 0,
  cache_ttl = NULL, key = NULL, ...)

Arguments

taxonconceptID

The taxonconceptID (numeric), which is also the page number.

iucn

Include the IUCN Red List status object (Default: FALSE)

images

Limits the number of returned image objects (values 0 - 75)

videos

Limits the number of returned video objects (values 0 - 75)

sounds

Limits the number of returned sound objects (values 0 - 75)

maps

Limits the number of returned map objects (values 0 - 75)

text

Limits the number of returned text objects (values 0 - 75)

subjects

'overview' (default) to return the overview text (if exists), a pipe | delimited list of subject names from the list of EOL accepted subjects (e.g. TaxonBiology, FossilHistory), or 'all' to get text in any subject. Always returns an overview text as a first result (if one exists in the given context).

licenses

A pipe | delimited list of licenses or 'all' (default) to get objects under any license. Licenses abbreviated cc- are all Creative Commons licenses. Visit their site for more information on the various licenses they offer.

details

Include all metadata for data objects. (Default: FALSE)

common_names

Return all common names for the page's taxon (Default: FALSE)

synonyms

Return all synonyms for the page's taxon (Default: FALSE)

references

Return all references for the page's taxon (Default: FALSE)

taxonomy

(logical) Whether to return any taxonomy details from different taxon hierarchy providers, in an array named taxonconcepts (Default: TRUE)

vetted

If 'vetted' is given a value of '1', then only trusted content will be returned. If 'vetted' is '2', then only trusted and unreviewed content will be returned (untrusted content will not be returned). The default is to return all content. (Default: FALSE)

cache_ttl

The number of seconds you wish to have the response cached.

key

Your EOL API key; loads from .Rprofile, or you can specify the key manually the in the function call.

...

Curl options passed on to GET

Value

JSON list object, or data.frame.

Details

It's possible to return JSON or XML with the EOL API. However, this function only returns JSON for now.

Examples

## Not run: ------------------------------------
# (pageid <- eol_search('Pomatomus')$pageid[1])
# eol_pages(taxonconceptID=pageid)$scinames
## ---------------------------------------------