Get the GBIF backbone taxon ID from taxonomic names.

Get the GBIF backbone taxon ID from taxonomic names.

get_gbifid(sciname, ask = TRUE, verbose = TRUE, rows = NA,
  phylum = NULL, class = NULL, order = NULL, family = NULL,
  rank = NULL, method = "backbone", ...)

as.gbifid(x, check = FALSE)

# S3 method for gbifid
as.gbifid(x, check = FALSE)

# S3 method for character
as.gbifid(x, check = TRUE)

# S3 method for list
as.gbifid(x, check = TRUE)

# S3 method for numeric
as.gbifid(x, check = TRUE)

# S3 method for data.frame
as.gbifid(x, check = TRUE)

# S3 method for gbifid
as.data.frame(x, ...)

get_gbifid_(sciname, verbose = TRUE, rows = NA, method = "backbone")

Arguments

sciname

character; scientific name.

ask

logical; should get_colid be run in interactive mode? If TRUE and more than one ID is found for the species, the user is asked for input. If FALSE NA is returned for multiple matches.

verbose

logical; If TRUE the actual taxon queried is printed on the console.

rows

numeric; Any number from 1 to infinity. If the default NA, all rows are considered. Note that this function still only gives back a gbifid class object with one to many identifiers. See get_gbifid_ to get back all, or a subset, of the raw data that you are presented during the ask process.

phylum

(character) A phylum (aka division) name. Optional. See Filtering below.

class

(character) A class name. Optional. See Filtering below.

order

(character) An order name. Optional. See Filtering below.

family

(character) A family name. Optional. See Filtering below.

rank

(character) A taxonomic rank name. See rank_ref for possible options. Though note that some data sources use atypical ranks, so inspect the data itself for options. Optional. See Filtering below.

method

(character) one of "backbone" or "lookup". See Details.

...

Ignored

x

Input to as.gbifid

check

logical; Check if ID matches any existing on the DB, only used in as.gbifid

Value

A vector of taxonomic identifiers as an S3 class. If a taxon is not found an NA is given. If more than one identifier is found the function asks for user input if ask = TRUE, otherwise returns NA. If ask=FALSE and rows does not equal NA, then a data.frame is given back, but not of the uid class, which you can't pass on to other functions as you normally can.

Comes with the following attributes:

match (character) - the reason for NA, either 'not found', 'found' or if ask = FALSE then 'NA due to ask=FALSE')

multiple_matches (logical) - Whether multiple matches were returned by the data source. This can be TRUE, even if you get 1 name back because we try to pattern match the name to see if there's any direct matches. So sometimes this attribute is TRUE, as well as pattern_match, which then returns 1 resulting name without user prompt.

pattern_match (logical) - Whether a pattern match was made. If TRUE then multiple_matches must be TRUE, and we found a perfect match to your name, ignoring case. If FALSE

Details

Internally in this function we use a function to search GBIF's taxonomy, and if we find an exact match we return the ID for that match. If there isn't an exact match we return the options to you to pick from.

method parameter

"backbone" uses the /species/match GBIF API route, matching against their backbone taxonomy. We turn on fuzzy matching by default, as the search without fuzzy against backbone is quite narrow. "lookup" uses the /species/search GBIF API route, doing a full text search of name usages covering scientific and vernacular named, species descriptions, distributions and the entire classification.

Filtering

The parameters phylum, class, order, family, and rank are not used in the search to the data provider, but are used in filtering the data down to a subset that is closer to the target you want. For all these parameters, you can use regex strings since we use grep internally to match. Filtering narrows down to the set that matches your query, and removes the rest.

See also

classification

Other taxonomic-ids: get_boldid, get_colid, get_eolid, get_ids, get_natservid, get_nbnid, get_tolid, get_tpsid, get_tsn, get_uid, get_wormsid

Examples

## Not run: ------------------------------------
# get_gbifid(sciname='Poa annua')
# get_gbifid(sciname='Pinus contorta')
# get_gbifid(sciname='Puma concolor')
# 
# # multiple names
# get_gbifid(c("Poa annua", "Pinus contorta"))
# 
# # specify rows to limit choices available
# get_gbifid(sciname='Pinus')
# get_gbifid(sciname='Pinus', rows=10)
# get_gbifid(sciname='Pinus', rows=1:3)
# 
# # When not found, NA given
# get_gbifid(sciname="uaudnadndj")
# get_gbifid(c("Chironomus riparius", "uaudnadndj"))
# 
# # Narrow down results to a division or rank, or both
# ## Satyrium example
# ### Results w/o narrowing
# get_gbifid("Satyrium")
# ### w/ phylum
# get_gbifid("Satyrium", phylum = "Tracheophyta")
# get_gbifid("Satyrium", phylum = "Arthropoda")
# ### w/ phylum & rank
# get_gbifid("Satyrium", phylum = "Arthropoda", rank = "genus")
# 
# ## Rank example
# get_gbifid("Poa", method = "lookup")
# get_gbifid("Poa", method = "lookup", rank = "genus")
# get_gbifid("Poa", method = "lookup", family = "Thripidae")
# 
# # Fuzzy filter on any filtering fields
# ## uses grep on the inside
# get_gbifid("Satyrium", phylum = "arthropoda")
# get_gbifid("A*", method = "lookup", order = "*tera")
# get_gbifid("A*", method = "lookup", order = "*ales")
# 
# # Convert a uid without class information to a uid class
# as.gbifid(get_gbifid("Poa annua")) # already a uid, returns the same
# as.gbifid(get_gbifid(c("Poa annua","Puma concolor"))) # same
# as.gbifid(2704179) # numeric
# as.gbifid(c(2704179,2435099,3171445)) # numeric vector, length > 1
# as.gbifid("2704179") # character
# as.gbifid(c("2704179","2435099","3171445")) # character vector, length > 1
# as.gbifid(list("2704179","2435099","3171445")) # list, either numeric or character
# ## dont check, much faster
# as.gbifid("2704179", check=FALSE)
# as.gbifid(2704179, check=FALSE)
# as.gbifid(2704179, check=FALSE)
# as.gbifid(c("2704179","2435099","3171445"), check=FALSE)
# as.gbifid(list("2704179","2435099","3171445"), check=FALSE)
# 
# (out <- as.gbifid(c(2704179,2435099,3171445)))
# data.frame(out)
# as.uid( data.frame(out) )
# 
# # Get all data back
# get_gbifid_("Puma concolor")
# get_gbifid_(c("Pinus", "uaudnadndj"))
# get_gbifid_(c("Pinus", "Puma"), rows=5)
# get_gbifid_(c("Pinus", "Puma"), rows=1:5)
# 
# # use curl options
# library("httr")
# get_gbifid("Quercus douglasii", config=verbose())
# bb <- get_gbifid("Quercus douglasii", config=progress())
## ---------------------------------------------