NCBI taxon information from uids

Downloads summary taxon information from the NCBI taxonomy databases for a set of taxonomy UIDs using eutils esummary.

ncbi_get_taxon_summary(id, ...)

Arguments

id

(character) NCBI taxonomy uids to retrieve information for. See Details.

...

Curl options passed on to GET

Value

A data.frame with the following columns:

uid

The uid queried for

name

The name of the taxon; a binomial name if the taxon is of rank species

rank

The taxonomic rank (e.g. 'Genus')

Details

If your input vector or list of NCBI IDs is longer than about 8000 characters (use nchar(paste(ids, collapse = "+"))), split the list up into chunks since at about that number of characters you will run into the HTTP 414 error "Request-URI Too Long".

Examples

## Not run: ------------------------------------
# ncbi_get_taxon_summary(c(1430660, 4751))
# 
# # use curl options
# library("httr")
# ncbi_get_taxon_summary(c(1430660, 4751), config = verbose())
## ---------------------------------------------