Retrieve synonyms from various sources given input taxonomic

Retrieve synonyms from various sources given input taxonomic names or identifiers

synonyms(...)

# S3 method for default
synonyms(x, db = NULL, rows = NaN, ...)

# S3 method for tsn
synonyms(id, ...)

# S3 method for colid
synonyms(id, ...)

# S3 method for tpsid
synonyms(id, ...)

# S3 method for nbnid
synonyms(id, ...)

# S3 method for wormsid
synonyms(id, ...)

# S3 method for ids
synonyms(id, ...)

synonyms_df(x)

Arguments

...

Other passed arguments to internal functions get_*() and functions to gather synonyms.

x

Vector of taxa names (character) or IDs (character or numeric) to query.

db

character; database to query. either itis, tropicos, col, nbn, or worms. Note that each taxonomic data source has their own identifiers, so that if you provide the wrong db value for the identifier you could get a result, but it will likely be wrong (not what you were expecting).

rows

(numeric) Any number from 1 to infinity. If the default NA, all rows are considered. Note that this parameter is ignored if you pass in a taxonomic id of any of the acceptable classes: tsn, tpsid, nbnid, ids.

id

character; identifiers, returned by get_tsn, get_tpsid, get_nbnid, get_colid, get_wormsid

Value

A named list of data.frames with the synonyms of every supplied taxa.

Details

If IDs are supplied directly (not from the get_* functions) you must specify the type of ID.

For db = "itis" you can pass in a parameter accepted to toggle whether only accepted names are used accepted = TRUE, or if all are used accepted = FALSE. The default is accepted = FALSE

See also

get_tsn, get_tpsid, get_nbnid, get_colid, get_wormsid

Examples

## Not run: ------------------------------------
# # Plug in taxon IDs
# synonyms(183327, db="itis")
# synonyms("25509881", db="tropicos")
# synonyms("NBNSYS0000004629", db='nbn')
# synonyms("87e986b0873f648711900866fa8abde7", db='col')
# synonyms(105706, db='worms')
# 
# # Plug in taxon names directly
# synonyms("Pinus contorta", db="itis")
# synonyms("Puma concolor", db="itis")
# synonyms(c("Poa annua",'Pinus contorta','Puma concolor'), db="itis")
# synonyms("Poa annua", db="tropicos")
# synonyms("Pinus contorta", db="tropicos")
# synonyms(c("Poa annua",'Pinus contorta'), db="tropicos")
# synonyms("Pinus sylvestris", db='nbn')
# synonyms("Puma concolor", db='col')
# synonyms("Ursus americanus", db='col')
# synonyms("Amblyomma rotundatum", db='col')
# synonyms('Pomatomus', db='worms')
# synonyms('Pomatomus saltatrix', db='worms')
# 
# # not accepted names, with ITIS
# ## looks for whether the name given is an accepted name,
# ## and if not, uses the accepted name to look for synonyms
# synonyms("Acer drummondii", db="itis")
# synonyms("Spinus pinus", db="itis")
# 
# # Use get_* methods
# synonyms(get_tsn("Poa annua"))
# synonyms(get_tpsid("Poa annua"))
# synonyms(get_nbnid("Carcharodon carcharias"))
# synonyms(get_colid("Ornithodoros lagophilus"))
# 
# # Pass many ids from class "ids"
# out <- get_ids(names="Poa annua", db = c('itis','tropicos'))
# synonyms(out)
# 
# # Use the rows parameter to select certain rows
# synonyms("Poa annua", db='tropicos', rows=1)
# synonyms("Poa annua", db='tropicos', rows=1:3)
# synonyms("Pinus sylvestris", db='nbn', rows=1:3)
# synonyms("Amblyomma rotundatum", db='col', rows=2)
# synonyms("Amblyomma rotundatum", db='col', rows=2:3)
# 
# # Use curl options
# synonyms("Poa annua", db='tropicos', rows=1, config=verbose())
# synonyms("Poa annua", db='itis', rows=1, config=verbose())
# synonyms("Poa annua", db='col', rows=1, config=verbose())
# 
# 
# # combine many outputs together
# x <- synonyms(c("Osmia bicornis", "Osmia rufa", "Osmia"), db = "itis")
# synonyms_df(x)
# 
# ## note here how Pinus contorta is dropped due to no synonyms found
# x <- synonyms(c("Poa annua",'Pinus contorta','Puma concolor'), db="col")
# synonyms_df(x)
# 
# ## note here that ids are taxon identifiers b/c you start with them
# x <- synonyms(c(25509881, 13100094), db="tropicos")
# synonyms_df(x)
# 
# ## xxx
# x <- synonyms(c('Aglais io', 'Usnea hirta', 'Arctostaphylos uva-ursi'),
#   db="nbn")
# synonyms_df(x)
## ---------------------------------------------