Defunct functions in taxize

The following functions are now defunct (no longer available):

Details

col_classification: See classification

eol_hierarchy: See classification

tp_classification: See classification

tpl_search: Use the Taxonstand functions TPL or TPLck directly.

get_seqs: This function changed name to ncbi_getbyname.

get_genes: This function changed name to ncbi_getbyid.

get_genes_avail: This function changed name to ncbi_search.

ncbi_getbyname: See ncbi_byname in the traits package.

ncbi_getbyid: See ncbi_byid in the traits package.

ncbi_search: See ncbi_searcher in the traits package.

eol_invasive: See eol in the originr package.

gisd_isinvasive: See gisd in the originr package.

ubio_classification: The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

ubio_classification_search: The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

ubio_id: The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

ubio_ping: The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

ubio_search: The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

ubio_synonyms: The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

get_ubioid: The uBio web services are apparently down indefinitely.

phylomatic_tree: This function is defunct. See phylomatic in the package brranching

phylomatic_format: This function is defunct. See phylomatic_names in the package brranching