Defunct functions in taxize

The following functions are now defunct (no longer available):

All COL functions are defunct: as.colid, col_children, col_classification, col_downstream, col_search, get_colid, get_colid_, as.data.frame.colid, children.colid, classification.colid, downstream.colid, id2name.colid, lowest_common.colid, synonyms.colid, upstream.colid`

col_classification(): Seeclassification()

tp_classification(): Seeclassification()

eol_hierarchy(): Seeclassification()

eol_invasive(): See eol in the originr package.

use_eol(): EOL no longer requires an API key

tpl_search(): Use the Taxonstand functions TPL or TPLck directly.

get_seqs(): This function changed name toncbi_getbyname()()].

get_genes(): This function changed name toncbi_getbyid()()].

get_genes_avail(): This function changed name toncbi_search()()].

ncbi_getbyname(): See ncbi_byname in the traits package.

ncbi_getbyid(): See ncbi_byid in the traits package.

ncbi_search(): See ncbi_searcher in the traits package.

gisd_isinvasive(): See gisd in the originr package.

ubio_classification(): The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

ubio_classification_search(): The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

ubio_id(): The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

ubio_ping(): The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

ubio_search(): The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

ubio_synonyms(): The uBio web services was down for quite a while, is now (as of 2016-05-09) back up, but we don't trust that it will stay up and available.

get_ubioid(): The uBio web services are apparently down indefinitely.

phylomatic_tree(): This function is defunct. See phylomatic in the package brranching

phylomatic_format(): This function is defunct. See phylomatic_names in the package brranching

eubon(): This function is defunct. Use eubon_search()

tnrs(): This function is defunct. Was too unreliable

tnrs_sources(): This function is defunct. Was too unreliable