Get citations and licenses for data sources used in taxize

Get citations and licenses for data sources used in taxize

taxize_cite(fxn = "itis", what = "citation")

Arguments

fxn

Function to search on. A special case is the package name 'taxize' that will give the citations for the package.

what

One of citation (default), license, or both.

Examples

taxize_cite(fxn='eol_search')

#> Source: eol 
#>   Home page: http://eol.org/ 
#>   API help: http://eol.org/api/ 
#>  
#> 
#> Can any of these citations be improved? https://github.com/ropensci/taxize/issues

taxize_cite(fxn='itis_hierarchy')

#> Source: itis 
#>   Home page: http://www.itis.gov/ 
#>   API help: http://www.itis.gov/ws_description.html 
#>   Citation: Retrieved [month, day, year], from the Integrated Taxonomic Information System on-line database, http://www.itis.gov. 
#> 
#> Can any of these citations be improved? https://github.com/ropensci/taxize/issues

taxize_cite(fxn='tp_classification')

#> Source: tropicos 
#>   Home page: http://tropicos.org/ 
#>   API help: http://services.tropicos.org/ 
#>  
#> 
#> Can any of these citations be improved? https://github.com/ropensci/taxize/issues

taxize_cite(fxn='gbif_ping')

#> Source: gbif 
#>   Home page: http://www.gbif.org 
#>   API help: http://www.gbif.org/developer/summary 
#>   Citation: GBIF (2013). GBIF (Ed.), Global Biodiversity Information Facility Data Portal (2013) 
#> 
#> Can any of these citations be improved? https://github.com/ropensci/taxize/issues

taxize_cite(fxn='plantminer')

#> Source: plantminer 
#>   Home page: http://www.plantminer.com/ 
#>   API help: http://www.plantminer.com/help 
#>   Citation: See The Plant List or Tropicos citations 
#> 
#> Can any of these citations be improved? https://github.com/ropensci/taxize/issues

taxize_cite(fxn='get_natservid_')

#> Source: natserv 
#>   Home page: http://www.natureserve.org/ 
#>   API help: https://services.natureserve.org/index.jsp 
#>   Citation: Citation: Natureserve. 2017. NatureServe Web Service. Arlington, VA. U.S.A. Available http://services.natureserve.org. (Accessed: <date>) 
#> 
#> Can any of these citations be improved? https://github.com/ropensci/taxize/issues

taxize_cite(fxn='as.natservid')

#> Source: natserv 
#>   Home page: http://www.natureserve.org/ 
#>   API help: https://services.natureserve.org/index.jsp 
#>   Citation: Citation: Natureserve. 2017. NatureServe Web Service. Arlington, VA. U.S.A. Available http://services.natureserve.org. (Accessed: <date>) 
#> 
#> Can any of these citations be improved? https://github.com/ropensci/taxize/issues

taxize_cite(fxn='get_wormsid')

#> Source: worms 
#>   Home page: http://www.marinespecies.org/ 
#>   API help: http://www.marinespecies.org/rest/ 
#>   Citation: We ask you to cite the individual global or regional species lists, or species pages as appropriate. Their citations are shown on their web pages. The database as a whole is to be cited as follows:
#> 
#>      WoRMS Editorial Board (2017). World Register of Marine Species. Available from http://www.marinespecies.org at VLIZ. Accessed <date>. doi:10.14284/170 
#> 
#> Can any of these citations be improved? https://github.com/ropensci/taxize/issues

taxize_cite(fxn='as.wormsid')

#> Source: worms 
#>   Home page: http://www.marinespecies.org/ 
#>   API help: http://www.marinespecies.org/rest/ 
#>   Citation: We ask you to cite the individual global or regional species lists, or species pages as appropriate. Their citations are shown on their web pages. The database as a whole is to be cited as follows:
#> 
#>      WoRMS Editorial Board (2017). World Register of Marine Species. Available from http://www.marinespecies.org at VLIZ. Accessed <date>. doi:10.14284/170 
#> 
#> Can any of these citations be improved? https://github.com/ropensci/taxize/issues

# Functions that use many data sources
taxize_cite(fxn='synonyms')

#> Source: itis 
#>   Home page: http://www.itis.gov/ 
#>   API help: http://www.itis.gov/ws_description.html 
#>   Citation: Retrieved [month, day, year], from the Integrated Taxonomic Information System on-line database, http://www.itis.gov. 
#> 
#> Source: col 
#>   Home page: http://www.catalogueoflife.org/ 
#>   API help: http://www.catalogueoflife.org/content/web-services 
#>   Citation: See > http://www.catalogueoflife.org/col/info/cite 
#> 
#> Source: tropicos 
#>   Home page: http://tropicos.org/ 
#>   API help: http://services.tropicos.org/ 
#>  
#> 
#> Source: nbn 
#>   Home page: http://www.nbn.org.uk/ 
#>   API help: https://data.nbn.org.uk/Documentation/Web_Services/ 
#>  
#> 
#> Source: worms 
#>   Home page: http://www.marinespecies.org/ 
#>   API help: http://www.marinespecies.org/rest/ 
#>   Citation: We ask you to cite the individual global or regional species lists, or species pages as appropriate. Their citations are shown on their web pages. The database as a whole is to be cited as follows:
#> 
#>      WoRMS Editorial Board (2017). World Register of Marine Species. Available from http://www.marinespecies.org at VLIZ. Accessed <date>. doi:10.14284/170 
#> 
#> Can any of these citations be improved? https://github.com/ropensci/taxize/issues

taxize_cite(fxn='classification')

#> Source: itis 
#>   Home page: http://www.itis.gov/ 
#>   API help: http://www.itis.gov/ws_description.html 
#>   Citation: Retrieved [month, day, year], from the Integrated Taxonomic Information System on-line database, http://www.itis.gov. 
#> 
#> Source: col 
#>   Home page: http://www.catalogueoflife.org/ 
#>   API help: http://www.catalogueoflife.org/content/web-services 
#>   Citation: See > http://www.catalogueoflife.org/col/info/cite 
#> 
#> Source: ncbi 
#>   Home page: http://www.ncbi.nlm.nih.gov/taxonomy 
#>   API help: http://www.ncbi.nlm.nih.gov/books/NBK25501/ 
#>   Citation: Federhen S: The NCBI Taxonomy database. Nucleic Acids Res 2012, 40 (Database issue): D136-D143. 
#> 
#> Source: gbif 
#>   Home page: http://www.gbif.org 
#>   API help: http://www.gbif.org/developer/summary 
#>   Citation: GBIF (2013). GBIF (Ed.), Global Biodiversity Information Facility Data Portal (2013) 
#> 
#> Source: eol 
#>   Home page: http://eol.org/ 
#>   API help: http://eol.org/api/ 
#>  
#> 
#> Source: troicos 
#>   Home page: http://tropicos.org/ 
#>   API help: http://services.tropicos.org/ 
#>  
#> 
#> Source: nbn 
#>   Home page: http://www.nbn.org.uk/ 
#>   API help: https://data.nbn.org.uk/Documentation/Web_Services/ 
#>  
#> 
#> Source: worms 
#>   Home page: http://www.marinespecies.org/ 
#>   API help: http://www.marinespecies.org/rest/ 
#>   Citation: We ask you to cite the individual global or regional species lists, or species pages as appropriate. Their citations are shown on their web pages. The database as a whole is to be cited as follows:
#> 
#>      WoRMS Editorial Board (2017). World Register of Marine Species. Available from http://www.marinespecies.org at VLIZ. Accessed <date>. doi:10.14284/170 
#> 
#> Source: natserv 
#>   Home page: http://www.natureserve.org/ 
#>   API help: https://services.natureserve.org/index.jsp 
#>   Citation: Citation: Natureserve. 2017. NatureServe Web Service. Arlington, VA. U.S.A. Available http://services.natureserve.org. (Accessed: <date>) 
#> 
#> Can any of these citations be improved? https://github.com/ropensci/taxize/issues

# Get the taxize citation
taxize_cite(fxn='taxize')

#> Warning: no date field in DESCRIPTION file of package ‘taxize’

#> Warning: could not determine year for ‘taxize’ from package DESCRIPTION file

#> The paper:  
#> 
#> To cite package ‘taxize’ in publications use:
#> 
#>   Scott Chamberlain, Eduard Szoecs and Zachary Foster (NA). taxize:
#>   Taxonomic Information from Around the Web. R package version
#>   0.8.4.9510. https://github.com/ropensci/taxize
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Manual{,
#>     title = {taxize: Taxonomic Information from Around the Web},
#>     author = {Scott Chamberlain and Eduard Szoecs and Zachary Foster},
#>     note = {R package version 0.8.4.9510},
#>     url = {https://github.com/ropensci/taxize},
#>   }
#> 
#> 
#> The software:  

#> Warning: subscript out of bounds

# Get license information
taxize_cite(fxn='taxize', "license")

#> License: MIT 
#> URL:     http://opensource.org/licenses/MIT